fix mongodb version and add exit to the server

+++ src/index.ts
@@ -1,7 +1,8 @@
 import express, { json } from "express";
-import { connect } from "./database";
+import { connect, disconnect } from "./database";
 import { IProduct, ProductModel } from "./models/ProductModel";
 import cors from "cors";
+import { Types } from "mongoose";
 
 const app = express();
 app.use(cors());
@@ -25,20 +26,50 @@ app.get("/products", async (req, res) => {
     }));
 });
 
+app.get("/product/:id", async (req, res) => {
+    if (!req.params.id || !Types.ObjectId.isValid(req.params.id)) {
+        res.sendStatus(404);
+        return;
+    }
+    const p = await ProductModel.findById(req.params.id);
+    if (!p) {
+        res.sendStatus(404);
+        return;
+    }
+    res.json({
+        id: p._id,
+        name: p.name,
+        desc: p.desc,
+        price: p.price
+    });
+});
+
 app.post("/product/add", async (req, res) => {
     const product = await ProductModel.create(req.body);
     res.json(product);
 });
 
 app.delete("/product/remove/:id", async (req, res) => {
-    if (req.params.id) {
+    if (req.params.id && Types.ObjectId.isValid(req.params.id)) {
         await ProductModel.findByIdAndRemove(req.params.id);
         res.json(req.params.id);
     } else {
-        res.status(404);
+        res.sendStatus(404);
     }
 });
 
-app.listen(3000, () => {
+const httpServer = app.listen(3000, () => {
     console.log("Server started at http://localhost:3000");
 });
+
+function handleShutdownGracefully() {
+    console.info("closing server gracefully...");
+    httpServer.close(() => {
+        console.info("server closed.");
+        disconnect();
+        process.exit(0);
+    });
+}
+process.on("SIGINT", handleShutdownGracefully);
+process.on("SIGTERM", handleShutdownGracefully);
+process.on("SIGHUP", handleShutdownGracefully);
